made it easier to run tests by running both scanning and ar marker checking at once
